DABOM-130 rename: sse를 관리하는 패키지를 Infra로 이전
