+++ src/main/java/com/project/domain/usagerecord/infra/messaging/UsageRecordKafkaConsumer.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Component;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.project.domain.usagerecord.UsageRecordService;
+import com.project.domain.usagerecord.service.UsageRecordService;
 import com.project.global.event.dto.EventEnvelope;
 import com.project.global.event.dto.usage.UsageRealtimePayload;
